@@ -12,26 +12,27 @@
     "clean": "rimraf wwwroot"
   },
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
+    "@angular/animations": "^6.0.3",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@ng-select/ng-select": "^2.1.3",
     "bootstrap-sass": "^3.3.7",
-    "core-js": "^2.5.5",
-    "ngx-bootstrap": "^2.0.3",
-    "rxjs": "5.5.10",
-    "sweetalert2": "^7.18.0",
-    "videogular2": "^6.0.8",
+    "ngx-bootstrap": "^3.0.0",
+    "rxjs": "^6.2.0",
+    "sweetalert2": "^7.21.1",
+    "videogular2": "^6.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angularclass/hmr": "^2.1.3",
     "@ngtools/webpack": "1.8.3",
+    "@types/jquery": "^3.3.2",
     "@types/node": "^8.0.54",
     "@types/webpack-env": "1.13.0",
     "angular-router-loader": "^0.8.0",
@@ -56,7 +57,7 @@
     "rimraf": "^2.6.2",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.19.0",
-    "typescript": "2.6.1",
+    "typescript": "2.9.1",
     "url-loader": "^0.6.2",
     "webpack": "^3.8.1",
     "webpack-bundle-analyzer": "^2.9.1",
